requirements.txt
lark-parser
pytest
autopep8
